:truck: Rename web/apis folder to web/api.
Why? Because this folder represents _the_ api mapping to the back-end, rather than a collection of apis to various locations.

If there was an api that corresponded to something else, such as auth0, then it would go in a folder named `web/auth0-api/`
