+++ web/src/api/README.md
@@ -1,12 +1,14 @@
-# APIs
+# API
+
+> Explanation of Naming: _the_ API mapping to the back-end, if there was an api that corresponded to something else, such as auth0, then it would go in a folder named `web/auth0-api/`
 
 APIs provide a way to map back-end endpoints to standard create/read/update/delete (CRUD) actions in the front-end. They handle api setup and data transformation. This lets them remain very simple and flexible if data needs to be transformed to match a new format from the backend, before the front-end gets updated to handle the new format.
 
 e.g.
 
 ```javascript
-// apis/forms-api.js
-import http from "@/apis/http-client"
+// api/forms-api.js
+import http from "@/api/http-client"
 
 export const formsApi = {
   create(attributes) {
