+++ web/src/api/expenses-api.js
@@ -1,4 +1,4 @@
-import http from "@/apis/http-client"
+import http from "@/api/http-client"
 
 // Must match types in src/api/models/expense.ts
 export const TYPES = Object.freeze({
