+++ web/src/modules/travel-authorizations/store.js
@@ -3,10 +3,10 @@ import { isString, upperFirst, pick } from "lodash"
 import { FORM_URL, LOOKUP_URL } from "@/urls"
 import { secureGet, securePost } from "@/store/jwt"
 
-import expensesApi from "@/apis/expenses-api"
-import locationsApi from "@/apis/locations-api"
-import travelAuthorizationsApi from "@/apis/travel-authorizations-api"
-import usersApi from "@/apis/users-api"
+import expensesApi from "@/api/expenses-api"
+import locationsApi from "@/api/locations-api"
+import travelAuthorizationsApi from "@/api/travel-authorizations-api"
+import usersApi from "@/api/users-api"
 
 const state = {
   departments: [],
