+++ web/src/modules/travelDesk/views/TravelRequest.vue
@@ -22,7 +22,7 @@
     import TravelerRequests from "./Requests/TravelerRequests.vue";
     import { TRAVEL_DESK_URL, PROFILE_URL} from "../../../urls";
     import { secureGet } from "../../../store/jwt";
-    import locationsApi from "@/apis/locations-api"
+    import locationsApi from "@/api/locations-api"
 
     export default {
